Adding support for work holidays. More work to be done.
Holidays which occur on the same month and day every year are visible to
the user. Holidays which occur n days before or after another defined
holiday are coded but not yet visible to the user (no interface yet).
Holidays which are on the nth xxxday (e.g., Thanksgiving on 4th Thursday
of November) are forthcoming.